resolved merge conflicts. moved weapons test to location on package.
